requirements.txt
python-dotenv==0.21.0
Requests==2.31.0
